develop 1.09.2020 a1
Исправил баг с инициализацией рабочего экрана ( обновление пути сервера происходило раньше чем отображался экран, изза чего сыпались эесепшены о несуществующем поле.)
Сделал сохранение файла на облаке.

+++ Server/src/main/java/com/geekbrains/krilov/serverNetty/DataHandler.java
@@ -2,12 +2,12 @@ package com.geekbrains.krilov.serverNetty;
 
 import com.geekbrains.krilov.ByteCommands;
 import com.geekbrains.krilov.FileInfo;
-import com.geekbrains.krilov.PackageHandler;
 import com.google.gson.Gson;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -17,14 +17,22 @@ import java.util.stream.Collectors;
 
 public class DataHandler extends ChannelInboundHandlerAdapter {
 
+    private enum State {
+        IDLE, RECEIVING_DATA
+    }
+
+    private final String SERVER_VAULT_CATALOG = "./vault/";
     private final String login;
-    private String currentDir;
-    private String homeDir;
+    private final String homeDir;
+    private State currentState = State.IDLE;
+
+    private String fileName;
+    private long fileLength;
+    private long receivedFileSize;
 
     public DataHandler(String login) {
         this.login = login;
-        this.homeDir = "./vault/" + login;
-        this.currentDir = homeDir;
+        this.homeDir = SERVER_VAULT_CATALOG + login + "/";
     }
 
     @Override
@@ -34,32 +42,44 @@ public class DataHandler extends ChannelInboundHandlerAdapter {
 
         while (buf.readableBytes() > 0) {
 
-            byte commandByte = buf.readByte();
 
-            switch (commandByte) {
-                case ByteCommands.GET_FILE_COMMAND:
-                    System.out.println("STATE: Receiving file from " + login);
-                    break;
-                case ByteCommands.SERVER_ROOT_PATH_COMMAND:
-                    System.out.println("STATE: Sending root path to " + login);
-                    sendRootPath(ctx);
-                    break;
-                case ByteCommands.SEND_FILE_COMMAND:
-                    System.out.println("STATE: Sending file to " + login);
-                    break;
-                case ByteCommands.DELETE_FILE_COMMAND:
-                    System.out.println("STATE: Deleting file by " + login);
-                    deleteFile(ctx, buf);
-                    break;
-                case ByteCommands.GET_SERVER_FILE_LIST_COMMAND:
-                    System.out.println("STATE: Sending fileList to " + login);
-                    sendFileList(ctx, buf);
-                    break;
-                default:
-                    System.out.println("unknown command byte");
-                    break;
+            if (currentState == State.IDLE) {
+
+                byte commandByte = buf.readByte();
+
+                switch (commandByte) {
+                    case ByteCommands.GET_FILE_COMMAND:
+                        System.out.println("STATE: Sending file to " + login);
+                        break;
+                    case ByteCommands.SERVER_ROOT_PATH_COMMAND:
+                        System.out.println("STATE: Sending root path to " + login);
+                        sendRootPath(ctx);
+                        break;
+                    case ByteCommands.SEND_FILE_COMMAND:
+                        System.out.println("STATE: Receiving file from " + login);
+                        getFile(ctx, buf);
+                        break;
+                    case ByteCommands.DELETE_FILE_COMMAND:
+                        System.out.println("STATE: Deleting file by " + login);
+                        deleteFile(ctx, buf);
+                        break;
+                    case ByteCommands.GET_SERVER_FILE_LIST_COMMAND:
+                        System.out.println("STATE: Sending fileList to " + login);
+                        sendFileList(ctx, buf);
+                        break;
+                    default:
+                        System.out.println("unknown command byte");
+                        break;
+                }
+            }
+
+            if (currentState == State.RECEIVING_DATA){
+                System.out.println("STATE: Receiving file data from " + login);
+                getFileData(ctx, buf);
             }
         }
+
+
     }
 
     @Override
@@ -115,12 +135,53 @@ public class DataHandler extends ChannelInboundHandlerAdapter {
     }
 
 
-    private void getFile(ChannelHandlerContext ctx, ByteBuf buf) throws Exception {
-        PackageHandler.receiveFile(buf, login);
+    private void sendFile(ChannelHandlerContext ctx, ByteBuf buf) throws Exception {
+
     }
 
-    private void sendFile(ChannelHandlerContext ctx, Object msg) {
-        PackageHandler.sendFile(msg, login);
+    private void getFile(ChannelHandlerContext ctx, ByteBuf buf) {
+        currentState = State.RECEIVING_DATA;
+        //получение длины имени файла
+        int fileNameLength = buf.readInt();
+        System.out.println("длина имени файла: " + fileNameLength);
+        //получение имени файла
+        byte[] fileNameBuf = new byte[fileNameLength];
+        buf.readBytes(fileNameBuf);
+        fileName = new String(fileNameBuf, StandardCharsets.UTF_8);
+        System.out.println("файл:" + fileName);
+        //получение длины файла
+        fileLength = buf.readLong();
+
+        if (fileLength > 0) {
+            receivedFileSize = 0;
+        } else receivedFileSize = -1L;
+
+        getFileData(ctx, buf);
+
+    }
+
+    private void getFileData(ChannelHandlerContext ctx, ByteBuf buf) {
+        //получение файла
+        boolean append = true;
+
+        try (FileOutputStream out = new FileOutputStream(homeDir + fileName, append)) {
+            System.out.println("получение файла");
+            while (buf.readableBytes() > 0) {
+                System.out.println("получено:" + receivedFileSize + " из " + fileLength);
+                int write = out.getChannel().write(buf.nioBuffer());
+                receivedFileSize += write;
+                buf.readerIndex(buf.readerIndex() + write);
+                if (receivedFileSize == fileLength) {
+                    ByteBuf tmp = Unpooled.buffer();
+                    tmp.writeByte(ByteCommands.OK_COMMAND);
+                    ctx.writeAndFlush(tmp);
+                    currentState = State.IDLE;
+                    break;
+                }
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     private String getFileListJson(String pathName) throws IOException {
